# c_openapi.yaml
openapi: 3.1.0
info:
  title: 住民記録システム Web版 API
  version: 0.1.0
  description: |
    住民記録システム標準仕様書 第6.1版（令和8年3月25日版 / 自治体システム等標準化検討会）に準拠した
    REST API の設計仕様（ドラフト）。ChromeOS Flex を含む Chromium 系ブラウザからの利用を想定。

    - すべての永続化操作は監査ログに記録される。
    - すべての応答は UTF-8 / application/json。日付は ISO 8601。
    - 認可は OAuth2 / OIDC のスコープ + 属性ベース（ロール + 項目別マスク）。
servers:
  - url: https://{tenant}.juki.example.go.jp/api/v1
    variables:
      tenant: { default: city01 }
...
security:
  - oidc: [residents.read, residents.write, transactions.write, certificates.issue, admin]
paths:
...
  /auth/login:
    post:
      tags: [auth]
      security: []
      summary: ログイン
      requestBody: { $ref: '#/components/requestBodies/LoginReq' }
      responses:
        '200': { $ref: '#/components/responses/LoginRes' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /residents/search:
    post:
      tags: [residents]
      summary: 住民検索
      description: 抑止対象は権限を持たないユーザには存在自体を隠蔽（404 相当を「該当なし」として返す）
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/ResidentSearchReq' } }
      responses:
        '200':
          description: OK
          content:
            application/json: { schema: { $ref: '#/components/schemas/PagedResidents' } }
...
  /transactions/in:
    post:
      tags: [transactions]
      summary: 転入届の受付・反映
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MoveInReq' } } }
      responses:
        '201':
          description: Created
          content: { application/json: { schema: { $ref: '#/components/schemas/Transaction' } } }
  /transactions/out:
    post:
      tags: [transactions]
      summary: 転出届
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MoveOutReq' } } } }
      responses:
        '201':
          description: Created（転出証明書発行情報を含む）
          content: { application/json: { schema: { $ref: '#/components/schemas/TransactionWithCertificate' } } }
...
  /transactions/cancel:
    post:
      tags: [transactions]
      summary: 異動取消（標準仕様 4.6）
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CancelReq' } } } }
      responses: { '201': { description: Created } }

  # 住民票コード / 個人番号 / 外国人在留
...
  /certificates/jumin:
    post:
      tags: [certificates]
      summary: 住民票の写し発行（0010001/0010003）
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CertificateReq' } } } }
      responses:
        '201':
          description: Created（PDFリンクと改ざん防止トークン）
          content: { application/json: { schema: { $ref: '#/components/schemas/CertificateIssue' } } }
...
  /restrictions:
    post:
      tags: [restrictions]
      summary: 抑止登録（DV等支援措置）
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Restriction' } } } }
      responses: { '201': { description: Created } }
...
  /reports/annual:
    post:
      tags: [reports]
      summary: 住基年報（標準仕様 20.6）
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ReportReq' } } } }
      responses:
        '202':
          description: 受付（非同期）
          content: { application/json: { schema: { $ref: '#/components/schemas/AsyncJob' } } }
...
  /euc/query:
    post:
      tags: [euc]
      summary: EUC（任意抽出）
      description: 個人番号が出力対象に含まれる場合は二段階承認、結果はパスワード付ZIP配信。
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/EucQueryReq' } } } }
      responses:
        '202': { description: 受付, content: { application/json: { schema: { $ref: '#/components/schemas/AsyncJob' } } } }

  # ===== 連携 =====
...
  /audit:
    get:
      tags: [admin]
      summary: 監査ログ参照
      parameters:
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to,   in: query, schema: { type: string, format: date-time } }
        - { name: userId, in: query, schema: { type: string } }
        - { name: residentId, in: query, schema: { type: string } }
        - { name: action, in: query, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json: { schema: { type: array, items: { $ref: '#/components/schemas/AuditLog' } } }
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://idp.example.go.jp/.well-known/openid-configuration
  requestBodies:
    LoginReq:
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              employeeId: { type: string }
              password:   { type: string, format: password }
              webauthn:   { type: object, description: WebAuthn アサーション（任意） }
  responses:
    LoginRes:
      description: OK
      content:
        application/json:
          schema:
            type: object
            properties:
              accessToken: { type: string }
              expiresIn:   { type: integer }
              me:          { $ref: '#/components/schemas/Me' }
    Unauthorized: { description: 認証エラー, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:    { type: string, example: VALIDATION_ERROR }
        message: { type: string }
        details: { type: array, items: { type: object } }
    Me:
      type: object
      properties:
        userId:     { type: string }
        fullName:   { type: string }
        department: { type: string }
        roles:      { type: array, items: { type: string } }
    Resident:
      type: object
      properties:
        residentId:        { type: string, description: 宛名番号 }
        familyNameKanji:   { type: string }
        givenNameKanji:    { type: string }
        familyNameKana:    { type: string }
        givenNameKana:     { type: string }
        birthDate:         { type: string, format: date }
        sex:               { type: string, enum: [M, F, U] }
        addressCode:       { type: string }
        addressText:       { type: string }
        householdId:       { type: string }
        relationToHead:    { type: string }
        movedInDate:       { type: string, format: date }
        movedOutDate:      { type: string, format: date, nullable: true }
        juminCode:         { type: string, description: 'マスク済 or 平文（権限による）' }
        myNumber:          { type: string, description: 'マスク済 or 平文（権限による）' }
        nationality:       { type: string, nullable: true }
        foreigner:         { $ref: '#/components/schemas/ForeignerInfo' }
        alias:             { type: array, items: { $ref: '#/components/schemas/AliasName' } }
        restrictions:      { type: array, items: { $ref: '#/components/schemas/Restriction' } }
        validFrom:         { type: string, format: date-time }
        validTo:           { type: string, format: date-time, nullable: true }
...
    ResidentSearchReq:
      type: object
      properties:
        name:        { type: string }
        kana:        { type: string }
        birthDate:   { type: string, format: date }
        sex:         { type: string, enum: [M, F] }
        residentId:  { type: string }
        juminCode:   { type: string }
        myNumber:    { type: string, description: 権限必須 }
        address:     { type: string }
        householdHeadOnly: { type: boolean }
        foreignerOnly:     { type: boolean }
        includeRemoved:    { type: boolean, default: false }
        page:        { type: integer, default: 1 }
        size:        { type: integer, default: 50 }
    PagedResidents:
      type: object
      properties:
        total: { type: integer }
        page:  { type: integer }
        size:  { type: integer }
        items: { type: array, items: { $ref: '#/components/schemas/Resident' } }
    ForeignerInfo:
      type: object
      properties:
        residenceStatus:        { type: string }
        residencePeriodEnd:     { type: string, format: date }
        passportNo:             { type: string }
        nationalityFull:        { type: string }
        aliasKanji:             { type: string, description: 通称 }
        specialPermanentResident: { type: boolean }
    AliasName:
      type: object
      properties:
        kind:      { type: string, enum: [ALIAS, FORMER_FAMILY] }
        valueKanji: { type: string }
        valueKana: { type: string }
        validFrom: { type: string, format: date }
        validTo:   { type: string, format: date, nullable: true }
    Restriction:
      type: object
      properties:
        id:          { type: string, readOnly: true }
        residentId:  { type: string }
        category:    { type: string, enum: [DV, STALKER, CHILD_ABUSE, OTHER] }
        startDate:   { type: string, format: date }
        endDate:     { type: string, format: date }
        scope:       { type: string, enum: [SELF, HOUSEHOLD] }
        releaseRole: { type: string }
        note:        { type: string }
    Transaction:
      type: object
      properties:
        transactionId:        { type: string }
        residentId:           { type: string }
        householdId:          { type: string }
        typeCode:             { type: string, description: 'IN/OUT/MOVE/HH/BIRTH/DEATH/OFFICIAL/CANCEL ...' }
        reasonCode:           { type: string }
        eventDate:            { type: string, format: date }
        processedDate:        { type: string, format: date }
        receiverOffice:       { type: string }
        status:               { type: string, enum: [DRAFT, REVIEW, APPROVED, APPLIED, CANCELLED] }
        parentTransactionId:  { type: string, nullable: true }
        items:                { type: array, items: { $ref: '#/components/schemas/TransactionItem' } }
    TransactionItem:
      type: object
      properties:
        field:       { type: string }
        valueBefore: { type: string, nullable: true }
        valueAfter:  { type: string, nullable: true }
    MoveInReq:
      type: object
      required: [eventDate, members]
      properties:
        eventDate:           { type: string, format: date }
        previousAddress:     { type: string }
        moveOutCertNo:       { type: string }
        members:             { type: array, items: { $ref: '#/components/schemas/MoveMember' } }
    MoveMember:
      type: object
      properties:
        familyNameKanji: { type: string }
        givenNameKanji:  { type: string }
        familyNameKana:  { type: string }
        givenNameKana:   { type: string }
        birthDate:       { type: string, format: date }
        sex:             { type: string, enum: [M, F, U] }
        relationToHead:  { type: string }
        myNumber:        { type: string }
        foreigner:       { $ref: '#/components/schemas/ForeignerInfo' }
    MoveOutReq:
      type: object
      required: [eventDate, members, newAddress]
      properties:
        eventDate:    { type: string, format: date }
        newAddress:   { type: string }
        newAddressCode: { type: string }
        reasonCode:   { type: string }
        members:      { type: array, items: { type: string, description: residentId } }
...
    CancelReq:
      type: object
      required: [transactionId, reason]
      properties:
        transactionId: { type: string }
        reason:        { type: string }
...
    CertificateReq:
      type: object
      required: [residentId, formId]
      properties:
        residentId:         { type: string }
        formId:             { type: string, enum: ["0010001", "0010002", "0010003", "0010004", "0010005", "0010007", "0010008"] }
        scope:              { type: string, enum: [SELF, HOUSEHOLD, MEMBERS] }
        showJuminCode:      { type: boolean }
        showMyNumber:       { type: boolean }
        showKoseki:         { type: boolean }
        showHeadAndRelation: { type: boolean }
        copies:             { type: integer }
        usageText:          { type: string }
    CertificateIssue:
      type: object
      properties:
        issueId:     { type: string }
        residentId:  { type: string }
        formId:      { type: string }
        copies:      { type: integer }
        fee:         { type: integer }
        verifyToken: { type: string }
        pdfUrl:      { type: string, format: uri }
        issuedAt:    { type: string, format: date-time }
        channel:     { type: string, enum: [WINDOW, CVS, ONLINE] }
...
    TransactionWithCertificate:
      allOf:
        - $ref: '#/components/schemas/Transaction'
        - type: object
          properties:
            certificate: { $ref: '#/components/schemas/CertificateIssue' }
    ReportReq:
      type: object
      properties:
        templateId: { type: string }
        fiscalYear: { type: integer }
        format:     { type: string, enum: [XLSX, PDF], default: XLSX }
        params:     { type: object }
    EucQueryReq:
      type: object
      properties:
        templateId: { type: string, nullable: true }
        sql:        { type: string, description: テンプレ ID なしの場合のみ }
        filters:    { type: object }
        outputFields: { type: array, items: { type: string } }
        format:     { type: string, enum: [CSV, XLSX], default: CSV }
        includeMyNumber: { type: boolean, default: false }
    AsyncJob:
      type: object
      properties:
        jobId:     { type: string }
        status:    { type: string, enum: [QUEUED, RUNNING, DONE, FAILED] }
        progress:  { type: integer, minimum: 0, maximum: 100 }
        resultUrl: { type: string, format: uri, nullable: true }
        error:     { type: string, nullable: true }
    AuditLog:
      type: object
      properties:
        logId:        { type: integer }
        userId:       { type: string }
        ip:           { type: string }
        action:       { type: string }
        resourceType: { type: string }
        resourceId:   { type: string }
        occurredAt:   { type: string, format: date-time }
        details:      { type: object, additionalProperties: true }